+++ Tests/TestGame.cs
@@ -36,6 +36,7 @@ public class TestGame : MlemGame {
 #endif
         ));
         // make sure we catch a potential ui stack overflow as part of the tests by ensuring a sufficient execution stack
+        this.UiSystem.OnElementAreaDirty += _ => RuntimeHelpers.EnsureSufficientExecutionStack();
         this.UiSystem.OnElementAreaUpdated += _ => RuntimeHelpers.EnsureSufficientExecutionStack();
     }